Organize all Express middlewares and functions into their own directories
