+++ graphql-server/src/schemas/neo4j.ts
@@ -4,7 +4,7 @@ import { OGM } from "@neo4j/graphql-ogm";
 import { toGraphQLTypeDefs } from "@neo4j/introspector";
 import { createHttpLink } from "apollo-link-http";
 import { InMemoryCache } from "apollo-cache-inmemory";
-import { buildProps } from "../buildProps";
+import { buildProps } from "../utils/buildProps";
 import { SamplesDocument, SortDirection } from "../generated/graphql";
 import { connect, headers, StringCodec } from "nats";
 const fetch = require("node-fetch");
